Some tweaks
Game is still in a rough prototype state.  Issues in ground check between ground and square colliders are causing bugs in movement and jumping. I think the problem comes from the scenario where you are touching the ground while a square passes through you.

Known bugs:

When a box touches the player for the first time, the player's RigidBody2D becomes the  child of the box, and starts being dragged with the box. The problem is that this is only supposed to happen when the player is standing on top of a box's platform effector collider. What's happening in the draft is that if you touch any box, the dragging effect happens for the rest of the game. I can either fix this bug completely so that the player remains stationary on the ground, or partially fix it so the game turns into an automatic side scroller.

Jumping is not consistent. Sometimes, the jump button doesn't respond when  pressed, and the issue happens usually after I pass through a passing box. I think this issue is happening for the same reasons as the box-dragging bug. More specifically, the game doesn't know what to do if the player's collider is touching both a box and the floor at the same time. If you are trying to hide from the red light, then you WILL have to be on the floor and in front of a block at the same time.

Sprite animations aren't implemented. This is a problem because the entire mechanic of the game revolves around the background switching between five sprites; green_light, yellow_light, red_light, safe_for_now, and found_you. I also want my character to move. Every asset is already in unity, and just needs to be plugged in.