Red Light District v0.3
Player movement is fully implemented, and working as planned... mostly! I am now trying to code in an animation controller with a script that makes the player character move. If I can do that, then I can use other animation controllers to make the coin move, and most importantly, change the background! I need to be able to change the background for the game's "red light, green light" mechanic to actually work.

Known bugs:

When pressed against a border wall, holding both down and the horizontal key in the direction of the border wall makes the player spaz out in the direction of the wall. This doesn't detract from gameplay, but it could potentially fudge up animation behaviors if the game struggles to decide between two animation states, player_run and player_sus.