Red Light District v0.4 (FINALLY)
Okay, I ran into a major hiccup with GitHub, but I was finally able to fix it. NOW:

I animated the coin.

I got started with making the Animator for the background. It has 5 states, and they all have their corresponding sprites loaded correctly. They are set to use a trigger to transition, and for Red Light, its two transition branches are determined with a bool. We need to code this so that it runs off a series of timers, can pause the game, and will check if the player is behind a platform when it pauses.