Background Manager script!
We want this script to assign "backgroundTimer" to a random int or float between the min and max values on lines 12-13 and 16-17, and then tick down every second or every frame down to zero. When the timer reaches zero, change the background with the "phase_change" trigger I put on the background animation controller.