First build on Itch.io is live!
Unfixed issues:

No UI to show coin count. There is a variable that updates as coins are collected in inspector, but no UI for it.

No way to tell if player is hiding or not on red light game state.

No way to pause or stop the game if the player is not hiding. This means that there is no way to trigger Game Over.